requirements.txt
anyio==4.0.0
certifi==2023.7.22
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
python-telegram-bot==20.6
sniffio==1.3.0
